backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=22.0
whitenoise>=6.6
